Worked on both mains xml
Main: has the a real estate pictures to display what the person is looking at and an enter button. What the enter button does is that it allows it to go to the next activity.
Second main: the next activity area where it hold a image as the same as the main.